smart_aqi_backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
requests
numpy
pandas
scikit-learn
joblib